+++ sdk/FunctionMetadataLoaderExtension/FunctionMetadataJsonReader.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Extensions.Options;
@@ -32,38 +33,32 @@ namespace Microsoft.Azure.WebJobs.Extensions.FunctionMetadataLoader
                 throw new FileNotFoundException($"Function metadata file not found. File path used:{metadataFile}");
             }
 
-            using (var fs = File.OpenText(metadataFile))
+            using var fs = File.OpenText(metadataFile);
+            using var js = new JsonTextReader(fs);
+            JArray functionMetadataJson = (JArray)await JToken.ReadFromAsync(js);
+            var functionList = new List<FunctionMetadata>();
+
+            foreach (JToken function in functionMetadataJson)
             {
-                using (var js = new JsonTextReader(fs))
+                FunctionMetadata metadata = function.ToObject<FunctionMetadata>()!;
+
+                // We need to re-add these by going through the BindingMetadata factory
+                metadata.Bindings.Clear();
+
+                if (function["bindings"] is not JArray bindingArray || bindingArray.Count == 0)
                 {
-                    JArray functionMetadataJson = (JArray)await JToken.ReadFromAsync(js);
-
-                    var functionList = new List<FunctionMetadata>();
-
-                    foreach (JObject function in functionMetadataJson)
-                    {
-                        FunctionMetadata metadata = function.ToObject<FunctionMetadata>();
-
-                        // We need to re-add these by going through the BindingMetadata factory
-                        metadata.Bindings.Clear();
-
-                        JArray bindingArray = (JArray)function["bindings"];
-                        if (bindingArray == null || bindingArray.Count == 0)
-                        {
-                            throw new FormatException("At least one binding must be declared.");
-                        }
-
-                        foreach (JObject binding in bindingArray)
-                        {
-                            metadata.Bindings.Add(BindingMetadata.Create(binding));
-                        }
-
-                        functionList.Add(metadata);
-                    }
-
-                    return functionList.ToImmutableArray();
+                    throw new FormatException("At least one binding must be declared.");
                 }
+
+                foreach (JObject binding in bindingArray.Cast<JObject>())
+                {
+                    metadata.Bindings.Add(BindingMetadata.Create(binding));
+                }
+
+                functionList.Add(metadata);
             }
+
+            return functionList.ToImmutableArray();
         }
     }
 }