+++ test/Sdk.Generator.Tests/FunctionExecutor/NotGeneratedTests.cs
@@ -24,7 +24,6 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 typeof(FunctionAttribute).Assembly,
                 typeof(LoggingServiceCollectionExtensions).Assembly,
                 typeof(ServiceProviderServiceExtensions).Assembly,
-                typeof(ServiceCollection).Assembly,
                 typeof(ILogger).Assembly,
                 typeof(IConfiguration).Assembly,
                 typeof(HostBuilder).Assembly,