+++ extensions/Worker.Extensions.ServiceBus/release_notes.md
@@ -4,6 +4,7 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.ServiceBus <version>
+### Microsoft.Azure.Functions.Worker.Extensions.ServiceBus 5.17.0
 
-- <entry>
+- Added the `AutoCompleteMessages` property to the `ServiceBusTriggerAttribute`, enabling automatic completion
+to be configured for each function.