+++ extensions/Worker.Extensions.Tables/release_notes.md
@@ -4,6 +4,8 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.Tables <version>
+### Microsoft.Azure.Functions.Worker.Extensions.Tables 1.3.0
 
-- <entry>
+- Updating `Microsoft.Azure.WebJobs.Extensions.Tables` reference to 1.2.1
+- Updating `Azure.Data.Tables` dependency to 12.8.2
+- Updating `Microsoft.Extensions.Azure` dependency to 1.7.1